+++ packages/service-provider-server/package.json
@@ -40,6 +40,7 @@
   "dependencies": {
     "@mongosh/errors": "^0.0.8",
     "@mongosh/service-provider-core": "^0.0.8",
+    "@mongosh/shell-api": "^0.0.8",
     "@types/sinon": "^7.5.1",
     "@types/sinon-chai": "^3.2.3",
     "mongodb": "3.5.3 || ^3.5.5"
