+++ packages/shell-api/package.json
@@ -37,6 +37,7 @@
     "@mongosh/errors": "^0.0.8",
     "@mongosh/history": "^0.0.8",
     "@mongosh/i18n": "^0.0.8",
-    "@mongosh/service-provider-core": "^0.0.8"
+    "@mongosh/service-provider-core": "^0.0.8",
+    "bson": "^1.1.4"
   }
 }
