+++ packages/shell-api/src/collection.spec.ts
@@ -10,7 +10,6 @@ import AggregationCursor from './aggregation-cursor';
 import Explainable from './explainable';
 import { Cursor as ServiceProviderCursor, ServiceProvider } from '@mongosh/service-provider-core';
 import ShellInternalState from './shell-internal-state';
-import bson from 'bson';
 
 const sinonChai = require('sinon-chai'); // weird with import
 
@@ -64,7 +63,6 @@ describe('Collection', () => {
       bus = stubInterface<EventEmitter>();
       serviceProvider = stubInterface<ServiceProvider>();
       serviceProvider.initialDb = 'test';
-      serviceProvider.bsonLibrary = bson;
       internalState = new ShellInternalState(serviceProvider, bus);
       mongo = new Mongo(internalState);
       database = new Database(mongo, 'db1');
