+++ packages/shell-api/src/database.spec.ts
@@ -8,7 +8,6 @@ import Collection from './collection';
 import Mongo from './mongo';
 import { Cursor as ServiceProviderCursor, ServiceProvider } from '@mongosh/service-provider-core';
 import ShellInternalState from './shell-internal-state';
-import bson from 'bson';
 
 describe('Database', () => {
   describe('help', () => {
@@ -109,7 +108,6 @@ describe('Database', () => {
       bus = stubInterface<EventEmitter>();
       serviceProvider = stubInterface<ServiceProvider>();
       serviceProvider.initialDb = 'test';
-      serviceProvider.bsonLibrary = bson;
       internalState = new ShellInternalState(serviceProvider, bus);
       mongo = new Mongo(internalState);
       database = new Database(mongo, 'db1');
