+++ packages/shell-api/src/explainable.spec.ts
@@ -10,7 +10,6 @@ import Collection from './collection';
 import Explainable from './explainable';
 import { ServiceProvider } from '@mongosh/service-provider-core';
 import ShellInternalState from './shell-internal-state';
-import bson from 'bson';
 
 describe('Explainable', () => {
   describe('help', () => {
@@ -62,7 +61,6 @@ describe('Explainable', () => {
       bus = stubInterface<EventEmitter>();
       serviceProvider = stubInterface<ServiceProvider>();
       serviceProvider.initialDb = 'test';
-      serviceProvider.bsonLibrary = bson;
       internalState = new ShellInternalState(serviceProvider, bus);
       mongo = new Mongo(internalState);
       database = new Database(mongo, 'db1');
