+++ packages/shell-api/src/index.ts
@@ -10,6 +10,7 @@ import toIterator from './toIterator';
 import Shard from './shard';
 import ReplicaSet from './replica-set';
 import ShellApi from './shell-api';
+import modifyBson from './modify-bson';
 import {
   BulkWriteResult,
   CommandResult,
@@ -20,6 +21,7 @@ import {
   UpdateResult
 } from './result';
 import Mongo from './mongo';
+import ShellBson from './shell-bson';
 import {
   signatures,
   ShellResult
@@ -51,6 +53,7 @@ export {
   Mongo,
   Shard,
   ReplicaSet,
+  ShellBson,
   UpdateResult,
   toIterator,
   signatures,
@@ -61,5 +64,6 @@ export {
   Topologies,
   shellApiType,
   asShellResult,
-  ShellResult
+  ShellResult,
+  modifyBson
 };
