+++ packages/shell-api/src/shell-api.spec.ts
@@ -9,7 +9,6 @@ import Mongo from './mongo';
 import { ReplPlatform, ServiceProvider } from '@mongosh/service-provider-core';
 import { EventEmitter } from 'events';
 import ShellInternalState from './shell-internal-state';
-import bson from 'bson';
 
 describe('ShellApi', () => {
   describe('signatures', () => {
@@ -86,7 +85,6 @@ describe('ShellApi', () => {
       newSP.initialDb = 'test';
       serviceProvider = stubInterface<ServiceProvider>({ getNewConnection: newSP });
       serviceProvider.initialDb = 'test';
-      serviceProvider.bsonLibrary = bson;
       mongo = stubInterface<Mongo>();
       mongo._serviceProvider = serviceProvider;
       internalState = new ShellInternalState(serviceProvider, bus);
@@ -181,7 +179,6 @@ describe('ShellApi', () => {
       serviceProvider = stubInterface<ServiceProvider>({ getNewConnection: newSP });
       serviceProvider.initialDb = 'test';
       serviceProvider.platform = ReplPlatform.CLI;
-      serviceProvider.bsonLibrary = bson;
       mongo = stubInterface<Mongo>();
       mongo._serviceProvider = serviceProvider;
       internalState = new ShellInternalState(serviceProvider, bus);
