+++ packages/shell-api/src/shell-bson.spec.ts
@@ -1,15 +1,18 @@
 /* eslint @typescript-eslint/camelcase: 0, new-cap: 0 */
-import constructShellBson from './shell-bson';
+import shellBson from './shell-bson';
 import { expect } from 'chai';
 import { ALL_SERVER_VERSIONS, asShellResult } from './enums';
-import bsonLib from 'bson';
-const shellBson = constructShellBson(bsonLib);
+import modifyBson from './modify-bson';
+import bson from 'bson';
 
 const hex_1234 = '31323334';
 const b64_1234 = 'MTIzNA==';
 const utf_1234 = '1234';
 
 describe('Shell BSON', () => {
+  before(() => {
+    modifyBson(bson);
+  });
   describe('DBRef', () => {
     it('without new', () => {
       const s = shellBson.DBRef('namespace', 'oid');
